refactor(gym/types): split tests to new sibling tests_types_extra.rs (#1266)
Inline test module in types.rs is moved into NEW sibling tests_types_extra.rs.
Existing tests_types.rs (347 LOC) was not extended to avoid pushing it past 400.

- src/gym/types.rs: 468 → 246 LOC (under 400-line rule)
- src/gym/tests_types_extra.rs: new (~219 LOC)
- src/gym/mod.rs: register mod tests_types_extra
- super::* → super::types::* (sibling level)

Pure refactor: zero logic changes, all 13 tests pass.

Closes 10/70 of #1266 (this loop iteration), 19 PRs cumulative since #1285.
